refactor: 성공시 모달 닫기
