+++ src/app/group/[groupId]/problem-list/[id]/page.tsx
@@ -1,7 +1,7 @@
 import { getSolutionList } from "@/api/solution";
 import Sidebar from "@/common/component/Sidebar";
 import { sidebarWrapper } from "@/styles/shared.css";
-import ProblemSidebar from "@/view/group/index/ProblemSidebar";
+import ProblemSidebar from "@/view/group/problem-list/ProblemSidebar";
 import SolvedList from "@/view/group/problem-list/SolvedList";
 import { contentWrapper } from "@/view/group/problem-list/index.css";
