--- src/view/group/index/ProblemSidebar/index.css.ts
@@ -1,6 +1,0 @@
-import { style } from "@vanilla-extract/css";
-
-export const problemSidebarWrapper = style({
-  width: "100%",
-  padding: "4.8rem 1.5rem 0",
-});
